requirements.txt
opencv-python
numpy
mysql-connector-python